test util launch
