+++ harness/tests/launch/test_deepspeed.py
@@ -12,6 +12,7 @@ from deepspeed.launcher.runner import DEEPSPEED_ENVIRONMENT_NAME
 import determined as det
 import determined.launch.deepspeed
 from determined import constants, launch
+from tests.launch import test_util
 
 
 def test_parse_args() -> None:
@@ -64,7 +65,7 @@ def test_launch_multi_slot_chief(
     mock_cluster_info: mock.MagicMock,
     mock_subprocess: mock.MagicMock,
 ) -> None:
-    cluster_info = make_mock_cluster_info(["0.0.0.0", "0.0.0.1"], 0)
+    cluster_info = test_util.make_mock_cluster_info(["0.0.0.0", "0.0.0.1"], 0, 4)
     mock_cluster_info.return_value = cluster_info
     mock_start_time = time.time()
     mock_time.return_value = mock_start_time
@@ -93,7 +94,7 @@ def test_launch_multi_slot_chief(
 
     mock_subprocess.side_effect = mock_process
 
-    with set_resources_id_env_var():
+    with test_util.set_resources_id_env_var():
         launch.deepspeed.main(script)
 
     mock_cluster_info.assert_called_once()
@@ -135,7 +136,7 @@ def test_launch_multi_slot_fail(
     mock_cluster_info: mock.MagicMock,
     mock_subprocess: mock.MagicMock,
 ) -> None:
-    cluster_info = make_mock_cluster_info(["0.0.0.0", "0.0.0.1"], 0)
+    cluster_info = test_util.make_mock_cluster_info(["0.0.0.0", "0.0.0.1"], 0, 4)
     mock_cluster_info.return_value = cluster_info
     mock_start_time = time.time()
     mock_time.return_value = mock_start_time
@@ -166,7 +167,7 @@ def test_launch_multi_slot_fail(
 
     mock_subprocess.side_effect = mock_process
 
-    with set_resources_id_env_var():
+    with test_util.set_resources_id_env_var():
         with pytest.raises(ValueError, match="no sshd greeting"):
             launch.deepspeed.main(script)
 
@@ -198,7 +199,7 @@ def test_launch_multi_slot_fail(
 def test_launch_one_slot(
     mock_cluster_info: mock.MagicMock, mock_subprocess: mock.MagicMock
 ) -> None:
-    cluster_info = make_mock_cluster_info(["0.0.0.0"], 0)
+    cluster_info = test_util.make_mock_cluster_info(["0.0.0.0"], 0, 4)
     mock_cluster_info.return_value = cluster_info
     script = ["s1", "s2"]
     pid_server_cmd = launch.deepspeed.create_pid_server_cmd(
@@ -209,7 +210,7 @@ def test_launch_one_slot(
     log_redirect_cmd = launch.deepspeed.create_log_redirect_cmd()
     launch_cmd = pid_server_cmd + deepspeed_cmd + pid_client_cmd + log_redirect_cmd + script
 
-    with set_resources_id_env_var():
+    with test_util.set_resources_id_env_var():
         launch.deepspeed.main(script)
 
     mock_cluster_info.assert_called_once()
@@ -222,7 +223,7 @@ def test_launch_one_slot(
 @mock.patch("subprocess.Popen")
 @mock.patch("determined.get_cluster_info")
 def test_launch_fail(mock_cluster_info: mock.MagicMock, mock_subprocess: mock.MagicMock) -> None:
-    cluster_info = make_mock_cluster_info(["0.0.0.0"], 0)
+    cluster_info = test_util.make_mock_cluster_info(["0.0.0.0"], 0, 4)
     mock_cluster_info.return_value = cluster_info
     mock_subprocess.return_value.wait.return_value = 1
     script = ["s1", "s2"]
@@ -234,7 +235,7 @@ def test_launch_fail(mock_cluster_info: mock.MagicMock, mock_subprocess: mock.Ma
     log_redirect_cmd = launch.deepspeed.create_log_redirect_cmd()
     launch_cmd = pid_server_cmd + deepspeed_cmd + pid_client_cmd + log_redirect_cmd + script
 
-    with set_resources_id_env_var():
+    with test_util.set_resources_id_env_var():
         assert launch.deepspeed.main(script) == 1
 
     mock_cluster_info.assert_called_once()
@@ -250,9 +251,9 @@ def test_launch_fail(mock_cluster_info: mock.MagicMock, mock_subprocess: mock.Ma
 def test_launch_worker(
     mock_api: mock.MagicMock, mock_cluster_info: mock.MagicMock, mock_subprocess: mock.MagicMock
 ) -> None:
-    cluster_info = make_mock_cluster_info(["0.0.0.0", "0.0.0.1"], 1)
+    cluster_info = test_util.make_mock_cluster_info(["0.0.0.0", "0.0.0.1"], 1, 4)
     mock_cluster_info.return_value = cluster_info
-    with set_resources_id_env_var():
+    with test_util.set_resources_id_env_var():
         launch.deepspeed.main(["script"])
 
     mock_cluster_info.assert_called_once()
@@ -268,29 +269,3 @@ def test_launch_worker(
     expected_cmd = pid_server_cmd + sshd_cmd
     mock_subprocess.assert_called_once_with(expected_cmd)
 
-
-def make_mock_cluster_info(container_addrs: List[str], container_rank: int) -> det.ClusterInfo:
-    rendezvous_info_mock = det.RendezvousInfo(
-        container_addrs=container_addrs, container_rank=container_rank
-    )
-    cluster_info_mock = det.ClusterInfo(
-        master_url="localhost",
-        cluster_id="clusterId",
-        agent_id="agentId",
-        slot_ids=[0, 1, 2, 3],
-        task_id="taskId",
-        allocation_id="allocationId",
-        session_token="sessionToken",
-        task_type="TRIAL",
-        rendezvous_info=rendezvous_info_mock,
-    )
-    return cluster_info_mock
-
-
-@contextlib.contextmanager
-def set_resources_id_env_var() -> Iterator[None]:
-    try:
-        os.environ["DET_RESOURCES_ID"] = "containerId"
-        yield
-    finally:
-        del os.environ["DET_RESOURCES_ID"]
