+++ harness/tests/launch/test_util.py
@@ -1,0 +1,49 @@
+import contextlib
+import os
+
+import determined as det
+from typing import List, Iterator
+from tests.experiment import utils
+
+
+def make_mock_cluster_info(
+        container_addrs: List[str], container_rank: int, num_slots: int
+) -> det.ClusterInfo:
+    config = utils.make_default_exp_config({}, 100, "loss", None)
+    trial_info_mock = det.TrialInfo(
+        trial_id=1,
+        experiment_id=1,
+        trial_seed=0,
+        hparams={},
+        config=config,
+        latest_batch=0,
+        trial_run_id=0,
+        debug=False,
+        unique_port_offset=0,
+        inter_node_network_interface=None,
+    )
+    rendezvous_info_mock = det.RendezvousInfo(
+        container_addrs=container_addrs, container_rank=container_rank
+    )
+    cluster_info_mock = det.ClusterInfo(
+        master_url="localhost",
+        cluster_id="clusterId",
+        agent_id="agentId",
+        slot_ids=list(range(num_slots)),
+        task_id="taskId",
+        allocation_id="allocationId",
+        session_token="sessionToken",
+        task_type="TRIAL",
+        rendezvous_info=rendezvous_info_mock,
+        trial_info=trial_info_mock,
+    )
+    return cluster_info_mock
+
+
+@contextlib.contextmanager
+def set_resources_id_env_var() -> Iterator[None]:
+    try:
+        os.environ["DET_RESOURCES_ID"] = "resourcesId"
+        yield
+    finally:
+        del os.environ["DET_RESOURCES_ID"]
